+++ src/command/ECommand.ts
@@ -2,7 +2,7 @@ export const enum ECommand {
     CmdCodeActionAddInclude = 'ahk.nekoHelp.CmdCodeActionAddInclude',
     CmdCodeActionRenameInclude = 'ahk.nekoHelp.CmdCodeActionRenameInclude',
     CmdFindClassRef = 'ahk.nekoHelp.CmdFindClassRef',
-    CmdFindComObjConnectRegister = 'ahk.nekoHelp.CmdFindComObjConnectRegister',
+    CmdJustGotoLoc = 'ahk.nekoHelp.CmdFindComObjConnectRegister',
     CmdFindFuncRef = 'ahk.nekoHelp.CmdFindFuncRef',
     CmdFindLabelRef = 'ahk.nekoHelp.CmdFindLabelRef',
     CmdFindMethodRef = 'ahk.nekoHelp.CmdFindMethodRef',